Rest api working.

+++ src/main/java/pl/fastus/waluta/model/Rate.java
@@ -8,6 +8,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
+import java.math.BigDecimal;
 
 @NoArgsConstructor
 @Data
@@ -20,13 +21,13 @@ public class Rate {
 
     private String currency;
     private String code;
-    private Double mid;
+    private BigDecimal mid;
 
     @ManyToOne
     private Currencies currencies;
 
     @Builder
-    public Rate(String currency, String code, Double mid) {
+    public Rate(String currency, String code, BigDecimal mid) {
         this.currency = currency;
         this.code = code;
         this.mid = mid;
